WIP: split serve_handlers.go per-tool (185-line coverage debt unresolved)
Mechanical move: 1374 → 182 LOC on serve_handlers.go; 10 new files
cmd/serve_handler_<tool>.go, one per MCP tool factory. All build/vet/lint
checks green. task check exit 0. go test -race -count=2 ./cmd/... clean.
Test files NOT touched — handlers referenced by package-level name only.

Why this is WIP and not a merge candidate:
- coverage-gate flagged 185 new prod lines uncovered. Verified pre-existing
  on trunk (zero-hit ranges identical when the same coverage profile is
  taken without this branch's changes). The gate lacks rename/similarity
  detection; it treats every new file's contents as net-new code.
- Two paths to exit 0, neither pure mechanical move:
  1. Annotate ~185 lines with // coverage:ignore (obscures real debt)
  2. Backfill 185 lines of tests (multi-PR effort; needs fake daemon
     plumbing + error injection on multiple seams)

Preserved here as the starting point for a fresh `feat/godfile-decomposition`
campaign off post-merge main. The right order for that campaign:
  1. Pay down the 185-line test debt FIRST (on the unmoved files)
  2. THEN do this mechanical split (coverage-clean, no annotation churn)

No co-author. Pure preservation commit.
